Refactor Direct Connect
